Move GreedyBestFitScanner to sonar package
